+++ ethercrab-wire-derive/src/parse_struct.rs
@@ -1,6 +1,6 @@
 use crate::help::{all_valid_attrs, attr_exists, bit_width_attr, has_repr_packed, usize_attr};
 use std::ops::Range;
-use syn::{DataStruct, DeriveInput, Fields, FieldsNamed, Ident, Type, Visibility};
+use syn::{DataStruct, DeriveInput, Fields, FieldsNamed, Generics, Ident, Type, Visibility};
 
 #[derive(Clone)]
 pub struct StructMeta {
@@ -9,6 +9,7 @@ pub struct StructMeta {
     pub repr_packed: bool,
 
     pub fields: Vec<FieldMeta>,
+    pub generics: Generics,
 }
 
 #[derive(Clone)]
@@ -43,7 +44,12 @@ pub struct FieldMeta {
 
 pub fn parse_struct(
     s: DataStruct,
-    DeriveInput { attrs, ident, .. }: DeriveInput,
+    DeriveInput {
+        attrs,
+        ident,
+        generics,
+        ..
+    }: DeriveInput,
 ) -> syn::Result<StructMeta> {
     // --- Struct attributes
 
@@ -187,5 +193,6 @@ pub fn parse_struct(
         repr_packed,
         width_bits: width,
         fields: field_meta,
+        generics,
     })
 }