+++ ethercrab-wire-derive/CHANGELOG.md
@@ -15,6 +15,8 @@ Derives for `ethercrab`.
 
 - **(breaking)** [#230](https://github.com/ethercrab-rs/ethercrab/pull/230) Increase MSRV from 1.77
   to 1.79.
+- [#321](https://github.com/ethercrab-rs/ethercrab/pull/321) (@fpdotmonkey) Support limited generic
+  parameters for structs.
 
 ## [0.2.0] - 2024-07-28