+++ ethercrab-wire-derive/CHANGELOG.md
@@ -6,6 +6,8 @@ Derives for `ethercrab`.
 
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2026-03-16
+
 ## Added
 
 - [#343](https://github.com/ethercrab-rs/ethercrab/pull/343) (@marcfir) Added support for
@@ -53,8 +55,9 @@ Derives for `ethercrab`.
 - Initial release
 
 <!-- next-url -->
+[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-wire-derive-v0.3.0...HEAD
 
-[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-wire-derive-v0.2.0...HEAD
+[0.3.0]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-wire-derive-v0.2.0...ethercrab-wire-derive-v0.3.0
 [0.2.0]:
   https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-wire-derive-v0.1.4...ethercrab-wire-derive-v0.2.0
 [0.1.4]: